Add more tests for UDP and crypto
